Change whats allowed in.
The rules that were changed were allowing connections in on those ports.

 This setup isn't providing any public services therefore does not need to offer any open ports. 

I did included a single rule to allow a connection for ssh if you so desire. I do not recommend, because it can be a point of weakness and allow an attacker to gain control of a point of entry into your network and  therefore other machines.